updated the feed scoring method

+++ PetZania/Back-End/adoption-and-breeding-module/src/main/java/com/example/adoption_and_breeding_module/model/entity/PetPost.java
@@ -79,7 +79,7 @@ public class PetPost {
     private Instant updatedAt;
 
     @Transient
-    private double score;
+    private long score;
 
     @PrePersist
     public void onCreate() {
